+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 ## Version 0.12.2
 
 ### Changes
-+ Made the config button in the forge mod list work. 
++ Made the config button in the forge mod list work.
++ Blaze slayer pillar warning text is now much bigger.
++ Hides the new 'you earn x event exp' message.
+
+### Fixes
++ Fixed a bug that caused the right blaze slayer dagger highlight to show wrong values.
 
 ## Version 0.12.1